feat: add overwatch bug prediction gh workflow (#342)
Add overwatch bug predictor github workflow to this repo to enable the
alpha feature.

Note that this feature is currently being merged with the AI PR review
feature (work tracked
[here](https://linear.app/getsentry/project/25q2-error-prediction-becomes-a-ai-pr-review-feature-60c79fb6bcf8/issues)).
Once that project is completed, this github action workflow can be
removed (this cleanup task is tracked
[here](https://linear.app/getsentry/issue/PREVENT-168)).

Also note I created a token and added a secret to this repo to enable
this feature (`OVERWATCH_SENTRY_AUTH_TOKEN`). This can also get cleaned
up with above when ready.