chore: Remove overwatch CLI in favor of per-commit (#409)
With this [PR](https://github.com/codecov/overwatch/pull/449/files), I
added `getsentry/sentry-mcp` to the feature-flagged repos which are
configured to have the overwatch bug predictor run on every commit.

As a result, we no longer need this stopgap that triggered the overwatch
CLI. This PR removes/cleans up the old github action.

Here's an
[example](https://github.com/getsentry/sentry-mcp/pull/408#issuecomment-3091358730)
of what the comment should look like now.

Closes https://linear.app/getsentry/issue/PREVENT-187/sentry-mcp-repo